fix signed/unsigned issues

+++ xpcom/glue/nsQuickSort.cpp
@@ -170,11 +170,11 @@ loop:	SWAPINIT(a, es);
 	pn = (char *)a + n * es;
 	r = PR_MIN(pa - (char *)a, pb - pa);
 	vecswap(a, pb - r, r);
-	r = PR_MIN(pd - pc, pn - pd - es);
+	r = PR_MIN(pd - pc, (int)(pn - pd - es));
 	vecswap(pb, pn - r, r);
-	if ((r = pb - pa) > es)
+	if ((r = pb - pa) > (int)es)
         NS_QuickSort(a, r / es, es, cmp, data);
-	if ((r = pd - pc) > es) {
+	if ((r = pd - pc) > (int)es) {
 		/* Iterate rather than recurse to save stack space */
 		a = pn - r;
 		n = r / es;
